feat: add support for `ThrowsAsync`
